Build: Use AX_GCC_FUNC_ATTRIBUTE to detect __attribute__ support.
This way -Werror isn't needed for these checks as
AX_GCC_FUNC_ATTRIBUTE looks at compiler's warning messages to see
if an unsupported attribute was present. Both methods have their
own problems but I'm hoping that this is the better method.

It is known that the check for the ifunc attribute this doesn't
work with CompCert 3.12. It silently ignores unknown attributes
in function declarations and in static functions if the function
appears unused (even though the unknown attribute would make it used).